# Bugfix: quote label name in matchers when needed