Yet another issue I discovered fuzzying.

`{"0"="0"}` is now a valid parseable vector selector, but `parser.VectorSelector.String()` of that is `{0="0"}`, which is not.

This fixes the `Matcher.String()` method to quote the name _when needed_.

Obviously, _when needed_ came at a cost, so I also optimized `Match.String()` getting rid of the _expensive_ `fmt.Sprintf()` and making the net cost of the method cheaper in most cases.

(Updated benchmarks)
```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/model/labels
                                                              │   main_mem    │             buffer_mem              │
                                                              │    sec/op     │   sec/op     vs base                │
Matcher_String/short_name_equal-12                              123.30n ± 10%   41.86n ± 0%  -66.05% (p=0.000 n=10)
Matcher_String/short_quoted_name_not_equal-12                   131.10n ± 14%   54.25n ± 0%  -58.62% (p=0.000 n=10)
Matcher_String/short_quoted_name_with_quotes_not_equal-12       125.10n ± 12%   60.01n ± 1%  -52.03% (p=0.000 n=10)
Matcher_String/short_name_value_with_quotes_equal-12            134.30n ± 16%   47.78n ± 1%  -64.42% (p=0.000 n=10)
Matcher_String/short_name_and_long_value_regexp-12               301.6n ± 14%   213.7n ± 1%  -29.16% (p=0.000 n=10)
Matcher_String/short_name_and_long_value_with_quotes_equal-12    323.0n ± 17%   226.9n ± 1%  -29.74% (p=0.000 n=10)
Matcher_String/long_name_regexp-12                              134.65n ± 10%   82.34n ± 1%  -38.85% (p=0.000 n=10)
Matcher_String/long_quoted_name_regexp-12                        132.0n ± 14%   232.4n ± 0%  +76.06% (p=0.000 n=10)
Matcher_String/long_name_and_long_value_regexp-12                316.3n ± 13%   263.4n ± 3%  -16.71% (p=0.000 n=10)
Matcher_String/long_quoted_name_and_long_value_regexp-12         309.5n ± 12%   412.7n ± 3%  +33.35% (p=0.000 n=10)
Matcher_String/mixed-12                                          212.9n ± 11%   172.5n ± 1%  -19.00% (p=0.000 n=10)
geomean                                                          187.0n         124.7n       -33.33%

                                                              │  main_mem  │             buffer_mem             │
                                                              │    B/op    │    B/op     vs base                │
Matcher_String/short_name_equal-12                              48.00 ± 0%   16.00 ± 0%  -66.67% (p=0.000 n=10)
Matcher_String/short_quoted_name_not_equal-12                   48.00 ± 0%   16.00 ± 0%  -66.67% (p=0.000 n=10)
Matcher_String/short_quoted_name_with_quotes_not_equal-12       48.00 ± 0%   16.00 ± 0%  -66.67% (p=0.000 n=10)
Matcher_String/short_name_value_with_quotes_equal-12            48.00 ± 0%   16.00 ± 0%  -66.67% (p=0.000 n=10)
Matcher_String/short_name_and_long_value_regexp-12              96.00 ± 0%   64.00 ± 0%  -33.33% (p=0.000 n=10)
Matcher_String/short_name_and_long_value_with_quotes_equal-12   96.00 ± 0%   64.00 ± 0%  -33.33% (p=0.000 n=10)
Matcher_String/long_name_regexp-12                              96.00 ± 0%   64.00 ± 0%  -33.33% (p=0.000 n=10)
Matcher_String/long_quoted_name_regexp-12                       96.00 ± 0%   64.00 ± 0%  -33.33% (p=0.000 n=10)
Matcher_String/long_name_and_long_value_regexp-12               144.0 ± 0%   112.0 ± 0%  -22.22% (p=0.000 n=10)
Matcher_String/long_quoted_name_and_long_value_regexp-12        144.0 ± 0%   112.0 ± 0%  -22.22% (p=0.000 n=10)
Matcher_String/mixed-12                                         86.00 ± 0%   54.00 ± 0%  -37.21% (p=0.000 n=10)
geomean                                                         79.52        42.14       -47.00%

                                                              │  main_mem  │             buffer_mem             │
                                                              │ allocs/op  │ allocs/op   vs base                │
Matcher_String/short_name_equal-12                              3.000 ± 0%   1.000 ± 0%  -66.67% (p=0.000 n=10)
Matcher_String/short_quoted_name_not_equal-12                   3.000 ± 0%   1.000 ± 0%  -66.67% (p=0.000 n=10)
Matcher_String/short_quoted_name_with_quotes_not_equal-12       3.000 ± 0%   1.000 ± 0%  -66.67% (p=0.000 n=10)
Matcher_String/short_name_value_with_quotes_equal-12            3.000 ± 0%   1.000 ± 0%  -66.67% (p=0.000 n=10)
Matcher_String/short_name_and_long_value_regexp-12              3.000 ± 0%   1.000 ± 0%  -66.67% (p=0.000 n=10)
Matcher_String/short_name_and_long_value_with_quotes_equal-12   3.000 ± 0%   1.000 ± 0%  -66.67% (p=0.000 n=10)
Matcher_String/long_name_regexp-12                              3.000 ± 0%   1.000 ± 0%  -66.67% (p=0.000 n=10)
Matcher_String/long_quoted_name_regexp-12                       3.000 ± 0%   1.000 ± 0%  -66.67% (p=0.000 n=10)
Matcher_String/long_name_and_long_value_regexp-12               3.000 ± 0%   1.000 ± 0%  -66.67% (p=0.000 n=10)
Matcher_String/long_quoted_name_and_long_value_regexp-12        3.000 ± 0%   1.000 ± 0%  -66.67% (p=0.000 n=10)
Matcher_String/mixed-12                                         3.000 ± 0%   1.000 ± 0%  -66.67% (p=0.000 n=10)
geomean                                                         3.000        1.000       -66.67%
```